Add delta time to stale-while-revalidate header
As per specification (RFC 5861), the cache control parameter
stale-while-revalidate accepts a duration in seconds
on how long serving stale (but unvalidated) requests is
considered to be acceptable. As far as I can tell omitting
this duration is not allowed by RFC 5861 at all.

At least some cloud providers refuse to accept this header
when it omits a duration, while they accept it when containing
a duration.

As far as I can tell the next.js interpretation of leaving out a value
was that it means "infinite". This commit approximates an infinite
duration with 1 year, which is consistent with the maxage parameter
sent for requests without revalidation.