Add next-i18n-router as a resource for internationalized routing
Many developers are asking how to add internationalized routing with the App Router as it works with the Pages Router. The next-i18n-router library is a very helpful resource.

Unlike the example provided in these Next.js i18n docs, next-i18n-router does not require nesting all pages in a [lang] dynamic segment. It also allows for the default language to be accessible without a locale prefix in the path (just like in the Pages Router).

Also like the Pages Router, this library lets developers use the NEXT-LOCALE cookie to set a user's preferred language.