Refactor metadata og and twitter title to be always presented (#52320)
Follow up for #52196 

The og title should be always resolved as `AbsoluteTemplate`. Move all the title resolving inside the og metadata resolving, so that it's aligned with type. And we don't have to check title everywhere, when og/twitter metadata is resolved, `title` is always presented as a property with `AbsoluteTemplate` type

Closes NEXT-1399