Turbopack: Use a different chunking context for web entry (#52404)
This fixes Turbopack integration tests.

In https://github.com/vercel/next.js/pull/51928, I made it so all dev chunking contexts output to the same directory. However, in the case of the web entry source, this can lead to conflicts. This PR ensures the chunking context is different for that one case.