Implement new forking technique for vendored packages.
The original method of vendoring packages has a flaw where it turns bare specifiers such as `require('foo')` into path specifiers such as `require('/path/to/compiled/pacakge/foo')`. Unfortunately this makes main fields and export maps not work as expected and thus modules do not resolve correctly. To cover up this deficiency there was extra effort into mappings deeper files inside of packages but this does not scale well long term as it requires vigilance if the internal state of these packages ever changes.

The new approach tries to model a global replacement where the require and import statements don't change but the module resolver proxies the resolution to a new location while maintaining bare specifier semantics. This is accomplished with a modification to the require-hook which uses a custom path for require.resolve and a modification to the webpack config's resolver plugins to alias to the appropriate path globally.

Both methods can fork behavior on whether we are in app router context or not and can do other things like vary experimental channel for react packages.

This relies on storing these vendored packages in a node_modules folder so to keep this technique separate from general purpose precompilation we have a new folder `next/dist/vendored` which will how packages that need these new semantics. A counter-example would be the webpack package itself. This requires mapping a bunch of discrete files to precompiled alternatives where we don't really want to just replace a package in place. This will remain in the `next/dist/compiled` folder and continue to work as is

When preparing a next instance for tests we need to copy vendored/node_modules even though all other node_modules folders are omitted

Refactors part of next-trace-entrypoints-plugin to be more efficient and to allow for separate traces for App Router andPages Router. Also implements augmented module resolution for vendoring react and other packages while tracing.

Vendored node_modules were being deleted from packing unless I add them explicitly to files property of package.json
Also removed some extraneous code for forked files for nft since the new tracing plugin can pick the right react to trace