Yarn PnP posed a problem. the solution is to treat these vendored dependencies as optional dependencies. PnP will resolve them correctly but for any package manager that does not it won't matter because the require hook and related module resolvers will just reference them packages from the vendored dir.
This change also moved the vendored dir out of src/dist into /vendored. The reason is that these packages do not need aditional build time processing and it's better for the files to be checked in and also present in the packed package to avoid chicken/egg issues.

Additionally all vendored packages need a unique name now so the two additional packages client-only and server-only are renamed with -vendored suffixes