disable `colormin` feature from `cssnano`
cssnano by default tries to convert `rgb` colors to an `hsla` color when
the `hsla` color is shorter for minification purposes.

The issue with this is that the colors aren't the same and we are losing
data when converting to `hsla`.

Input:
```css
html {
  background-color: rgb(143 101 98 / 43%);
}
```

Output:
```css
html {
  background-color: hsla(4, 19%, 47%, .43);
}
```
(I pretty printed it to make the diff a bit nicer).

If we would convert this `hsla` color back to `rgb` syntax, then it
results in:
```css
html {
  background-color: rgb(143 100 97 / 43%);
}
```

Comparing the original value with the last value, then you can see the
difference:
```diff
  html {
-   background-color: rgb(143 101 98 / 43%);
+   background-color: rgb(143 100 97 / 43%);
  /*                            ^  ^        */
  }
```

To solve this issue, we can just disable the `colormin` minification of
`cssnano` entirely.